feat: Automatically infer the API and Hydra hostnames
For the cloud solution so that we can reuse the same Docker image of the Portal for all customers, even though all of them will be running on different hostnames.

When not running at `*.cloud.apisuite.io`, the API and Hydra hostnames continue to be read from the respective env vars.