chore: Rename the env var SUPPORT_URL to INFORM_URL
This env var is used to indicate the URL of an endpoint (in our case a lambda function in AWS) where the portal POSTs the email of users that register to know more about the product. Since in the cloud version the Support URL will be an URL configured in the Settings API of an URL to where we will redirect the user, it makes sense to rename this env var so that there is no confusion.