feat(Portal): New layouts-based layout.
Previously there was only a single layout. Now pages define which layout they want to be rendered in. There are two layouts: Main and Essential. The Main layout is the usual one with the navigation menu in the header. The Essential layout is for pages such as login where we don't want Navigation.